Finish merging JKCodeBuilder and JVisitorWithCommentsPrinting
