+++ plugins/kotlin/j2k/shared/src/org/jetbrains/kotlin/nj2k/printing/JKCodeBuilder.kt
@@ -10,7 +10,7 @@ import org.jetbrains.kotlin.nj2k.tree.*
 import org.jetbrains.kotlin.nj2k.tree.JKClass.ClassKind.*
 import org.jetbrains.kotlin.nj2k.tree.Modality.FINAL
 import org.jetbrains.kotlin.nj2k.tree.Visibility.PUBLIC
-import org.jetbrains.kotlin.nj2k.tree.visitors.JKVisitorWithCommentsPrinting
+import org.jetbrains.kotlin.nj2k.tree.visitors.JKVisitor
 import org.jetbrains.kotlin.nj2k.types.JKContextType
 import org.jetbrains.kotlin.nj2k.types.isAnnotationMethod
 import org.jetbrains.kotlin.nj2k.types.isInterface
@@ -27,12 +27,12 @@ class JKCodeBuilder(context: NewJ2kConverterContext) {
         return printer.toString().replace("\r\n", "\n")
     }
 
-    private inner class Visitor : JKVisitorWithCommentsPrinting() {
-        override fun printLeftNonCodeElements(element: JKFormattingOwner) {
+    private inner class Visitor : JKVisitor() {
+        fun printLeftNonCodeElements(element: JKFormattingOwner) {
             commentPrinter.printCommentsAndLineBreaksBefore(element)
         }
 
-        override fun printRightNonCodeElements(element: JKFormattingOwner) {
+        fun printRightNonCodeElements(element: JKFormattingOwner) {
             commentPrinter.printCommentsAndLineBreaksAfter(element)
         }
 
@@ -73,7 +73,156 @@ class JKCodeBuilder(context: NewJ2kConverterContext) {
             }
         }
 
-        override fun visitTreeElementRaw(treeElement: JKElement) {
+        override fun visitTreeElement(treeElement: JKElement) {
+            if (treeElement is JKFormattingOwner) {
+                printLeftNonCodeElements(treeElement)
+            }
+            visitTreeElementRaw(treeElement)
+            if (treeElement is JKFormattingOwner) {
+                printRightNonCodeElements(treeElement)
+            }
+        }
+
+        override fun visitDeclaration(declaration: JKDeclaration) {
+            printLeftNonCodeElements(declaration)
+            visitTreeElementRaw(declaration)
+            printRightNonCodeElements(declaration)
+        }
+
+        override fun visitVariable(variable: JKVariable) {
+            printLeftNonCodeElements(variable)
+            visitTreeElementRaw(variable)
+            printRightNonCodeElements(variable)
+        }
+
+        override fun visitJavaStaticInitDeclaration(javaStaticInitDeclaration: JKJavaStaticInitDeclaration) {
+            printLeftNonCodeElements(javaStaticInitDeclaration)
+            visitTreeElementRaw(javaStaticInitDeclaration)
+            printRightNonCodeElements(javaStaticInitDeclaration)
+        }
+
+        override fun visitLabel(label: JKLabel) {
+            printLeftNonCodeElements(label)
+            visitTreeElementRaw(label)
+            printRightNonCodeElements(label)
+        }
+
+
+        override fun visitKtWhenLabel(ktWhenLabel: JKKtWhenLabel) {
+            printLeftNonCodeElements(ktWhenLabel)
+            visitTreeElementRaw(ktWhenLabel)
+            printRightNonCodeElements(ktWhenLabel)
+        }
+
+        override fun visitJavaTryCatchSection(javaTryCatchSection: JKJavaTryCatchSection) {
+            printLeftNonCodeElements(javaTryCatchSection)
+            visitTreeElementRaw(javaTryCatchSection)
+            printRightNonCodeElements(javaTryCatchSection)
+        }
+
+        override fun visitJavaSwitchCase(javaSwitchCase: JKJavaSwitchCase) {
+            printLeftNonCodeElements(javaSwitchCase)
+            visitTreeElementRaw(javaSwitchCase)
+            printRightNonCodeElements(javaSwitchCase)
+        }
+
+        override fun visitJavaDefaultSwitchCase(javaDefaultSwitchCase: JKJavaDefaultSwitchCase) {
+            printLeftNonCodeElements(javaDefaultSwitchCase)
+            visitTreeElementRaw(javaDefaultSwitchCase)
+            printRightNonCodeElements(javaDefaultSwitchCase)
+        }
+
+        override fun visitJavaLabelSwitchCase(javaLabelSwitchCase: JKJavaLabelSwitchCase) {
+            printLeftNonCodeElements(javaLabelSwitchCase)
+            visitTreeElementRaw(javaLabelSwitchCase)
+            printRightNonCodeElements(javaLabelSwitchCase)
+        }
+
+        override fun visitExpression(expression: JKExpression) {
+            printLeftNonCodeElements(expression)
+            visitTreeElementRaw(expression)
+            printRightNonCodeElements(expression)
+        }
+
+        override fun visitOperatorExpression(operatorExpression: JKOperatorExpression) {
+            printLeftNonCodeElements(operatorExpression)
+            visitTreeElementRaw(operatorExpression)
+            printRightNonCodeElements(operatorExpression)
+        }
+
+        override fun visitUnaryExpression(unaryExpression: JKUnaryExpression) {
+            printLeftNonCodeElements(unaryExpression)
+            visitTreeElementRaw(unaryExpression)
+            printRightNonCodeElements(unaryExpression)
+        }
+
+        override fun visitKtAssignmentChainLink(ktAssignmentChainLink: JKKtAssignmentChainLink) {
+            printLeftNonCodeElements(ktAssignmentChainLink)
+            visitTreeElementRaw(ktAssignmentChainLink)
+            printRightNonCodeElements(ktAssignmentChainLink)
+        }
+
+        override fun visitJavaNewEmptyArray(javaNewEmptyArray: JKJavaNewEmptyArray) {
+            printLeftNonCodeElements(javaNewEmptyArray)
+            visitTreeElementRaw(javaNewEmptyArray)
+            printRightNonCodeElements(javaNewEmptyArray)
+        }
+
+        override fun visitJavaNewArray(javaNewArray: JKJavaNewArray) {
+            printLeftNonCodeElements(javaNewArray)
+            visitTreeElementRaw(javaNewArray)
+            printRightNonCodeElements(javaNewArray)
+        }
+
+        override fun visitJavaAssignmentExpression(javaAssignmentExpression: JKJavaAssignmentExpression) {
+            printLeftNonCodeElements(javaAssignmentExpression)
+            visitTreeElementRaw(javaAssignmentExpression)
+            printRightNonCodeElements(javaAssignmentExpression)
+        }
+
+        override fun visitStatement(statement: JKStatement) {
+            printLeftNonCodeElements(statement)
+            visitTreeElementRaw(statement)
+            printRightNonCodeElements(statement)
+        }
+
+        override fun visitLoopStatement(loopStatement: JKLoopStatement) {
+            printLeftNonCodeElements(loopStatement)
+            visitTreeElementRaw(loopStatement)
+            printRightNonCodeElements(loopStatement)
+        }
+
+        override fun visitJavaSwitchStatement(javaSwitchStatement: JKJavaSwitchStatement) {
+            printLeftNonCodeElements(javaSwitchStatement)
+            visitTreeElementRaw(javaSwitchStatement)
+            printRightNonCodeElements(javaSwitchStatement)
+        }
+
+        override fun visitJavaTryStatement(javaTryStatement: JKJavaTryStatement) {
+            printLeftNonCodeElements(javaTryStatement)
+            visitTreeElementRaw(javaTryStatement)
+            printRightNonCodeElements(javaTryStatement)
+        }
+
+        override fun visitJavaSynchronizedStatement(javaSynchronizedStatement: JKJavaSynchronizedStatement) {
+            printLeftNonCodeElements(javaSynchronizedStatement)
+            visitTreeElementRaw(javaSynchronizedStatement)
+            printRightNonCodeElements(javaSynchronizedStatement)
+        }
+
+        override fun visitJavaAssertStatement(javaAssertStatement: JKJavaAssertStatement) {
+            printLeftNonCodeElements(javaAssertStatement)
+            visitTreeElementRaw(javaAssertStatement)
+            printRightNonCodeElements(javaAssertStatement)
+        }
+
+        override fun visitJavaForLoopStatement(javaForLoopStatement: JKJavaForLoopStatement) {
+            printLeftNonCodeElements(javaForLoopStatement)
+            visitTreeElementRaw(javaForLoopStatement)
+            printRightNonCodeElements(javaForLoopStatement)
+        }
+
+        fun visitTreeElementRaw(treeElement: JKElement) {
             printer.print("/* !!! Hit visitElement for element type: ${treeElement::class} !!! */")
         }
